Move `.~/images/google-logo.svg` to `.~/images/logo`.
